+++ resources/js/Components/MaterialDesign/MdTextInput.vue
@@ -1,5 +1,5 @@
 <template>
-    <div class="relative w-full">
+    <div class="relative w-full my-3 px-1">
         <!-- Label flotante -->
         <label v-if="label" :for="id"
             class="absolute text-sm  transition-all duration-300 ease-in-out px-1 pointer-events-none z-10 flex items-center gap-1 transform"
@@ -34,14 +34,14 @@
 
         <!-- Input principal -->
         <input :id="id" :type="type" :name="name" ref="inputRef"
-            class="w-full h-10 border rounded-xl text-gray-800 dark:text-gray-100 placeholder-transparent focus:outline-none focus:ring-2 transition-all duration-300 ease-in-out shadow-sm focus:shadow-md"
+            class="w-full h-10 border rounded-xl text-gray-800 dark:text-gray-100 placeholder-white focus:outline-none focus:ring-2 transition-all duration-300 ease-in-out shadow-sm focus:shadow-md"
             :class="{
                 'opacity-50 cursor-not-allowed': disabled || readonly,
                 'pl-10 pr-4': iconLeft || iconClass,
                 'px-4': !iconLeft && !iconClass,
                 'ring-1 ring-inset': success
-            }" :placeholder="placeholder || label" :value="internalValue" :disabled="disabled" :readonly="readonly"
-            :minlength="minlength" :maxlength="maxlength" @focus="onFocus" @blur="onBlur"
+            }" :placeholder="showRealPlaceholder ? placeholder : ' '" :value="internalValue" :disabled="disabled" :readonly="readonly"
+            :minlength="minlength" :maxlength="maxlength" @focus="onFocus" @blur="onBlur"   @keydown="onKeydown"
             @input="updateValue($event.target.value)" :style="{
                 backgroundColor,
                 borderColor,
@@ -53,16 +53,27 @@
             {{ errorText }}
         </div>
 
-        <!-- Contador de caracteres -->
-        <div v-if="showCharCounter" class="text-right text-xs px-1 mt-0.5" :class="charCountColor">
-            {{ internalValue.length }} / {{ maxlength }}
+        <!-- Ayuda izquierda + contador derecha -->
+        <div v-if="helper  || showCharCounter"
+            class="flex items-center justify-between text-xs px-1 mt-1 leading-tight">
+
+            <!-- Texto ejemplo o ayuda -->
+            <div v-if="helper" class="text-gray-400 ml-1">
+                {{ helper }}
+            </div>
+
+            <!-- Contador de caracteres -->
+            <div v-if="showCharCounter" :class="charCountColor">
+                {{ internalValue.length }} / {{ maxlength }}
+            </div>
+
         </div>
+
     </div>
 </template>
 
 <script setup>
-import { ref, watch, computed, useSlots, onMounted, onBeforeUnmount } from 'vue'
-import IMask from 'imask'
+import { ref, watch, computed, useSlots, onMounted } from 'vue'
 
 /* ======================= Props ========================== */
 const props = defineProps({
@@ -79,13 +90,17 @@ const props = defineProps({
     error: [Boolean, String, Array],
     success: { type: Boolean, default: false },
     iconClass: { type: String, default: '' },
-    mask: [String, Object],
     minlength: Number,
     maxlength: Number,
+    regex: RegExp,
     inputRestrict: {
         type: String,
         validator: (val) => ['letters', 'numbers', 'alphanumeric', 'none'].includes(val),
         default: 'none'
+    },
+    helper: {
+        type: String,
+        default: ''
     },
 })
 
@@ -95,30 +110,28 @@ const emit = defineEmits(['update:modelValue', 'focus', 'blur'])
 /* ======================= Refs ========================== */
 const isFocused = ref(false)
 const internalValue = ref(props.modelValue ?? '')
+const internalError = ref('')
 const isDark = ref(false)
 const slots = useSlots()
 const inputRef = ref(null)
-let maskRef = null
+const showRealPlaceholder = false
 
 /* ======================= Watch modelValue ========================== */
 watch(() => props.modelValue, (val) => {
     internalValue.value = val
-    if (maskRef && maskRef.value !== val) maskRef.value = val
 })
 
-/* ======================= Watch mask ========================== */
-watch(() => props.mask, (newMask) => {
-    if (inputRef.value) {
-        if (maskRef) maskRef.destroy()
+watch(internalValue, (val) => {
+    internalError.value = ''
 
-        maskRef = IMask(
-            inputRef.value,
-            typeof newMask === 'string' ? { mask: newMask } : newMask
-        )
-
-        maskRef.on('accept', () => {
-            emit('update:modelValue', maskRef.value)
-        })
+    if (props.required && !val) {
+        internalError.value = 'Este campo es obligatorio'
+    } else if (props.minlength && val.length < props.minlength) {
+        internalError.value = `Debe tener al menos ${props.minlength} caracteres`
+    } else if (props.maxlength && val.length > props.maxlength) {
+        internalError.value = `Debe tener máximo ${props.maxlength} caracteres`
+    } else if (props.regex && !props.regex.test(val)) {
+        internalError.value = 'El formato no es válido'
     }
 })
 
@@ -130,7 +143,7 @@ const backgroundColor = computed(() =>
 )
 
 const borderColor = computed(() => {
-    if (props.error) return 'var(--color-complement-2)'
+    if (props.error || internalError.value) return 'var(--color-complement-2)'
     if (props.success) return 'var(--color-primary-hover)'
     if (isFocused.value) return 'var(--color-primary)'
     return 'var(--color-primary-light)'
@@ -141,18 +154,20 @@ const labelColor = computed(() =>
 )
 
 const errorText = computed(() => {
+    if (internalError.value) return internalError.value
     if (Array.isArray(props.error)) return props.error[0]
     if (typeof props.error === 'string') return props.error
     return null
 })
 
 const charCountColor = computed(() => {
-    const length = internalValue.value?.length || 0
-    const max = props.maxlength
-    if (!max) return 'text-gray-500'
-    const percentage = (length / max) * 100
-    if (percentage >= 100) return 'text-[var(--color-complement-2)]'
-    if (percentage >= 90) return 'text-orange-500'
+    const currentLength = internalValue.value?.length || 0
+    const maxLength = props.maxlength
+    if (!maxLength) return 'text-gray-500'
+
+    const usedPercentage = (currentLength / maxLength) * 100
+    if (currentLength > maxLength) return 'text-[var(--color-complement-2)]'
+    if (usedPercentage >= 90) return 'text-[var(--color-primary)]'
     return 'text-gray-500'
 })
 
@@ -164,18 +179,11 @@ const showCharCounter = computed(() =>
 function updateValue(val) {
     let finalValue = val
 
-    // Aplica restricciones si están habilitadas
     if (props.inputRestrict === 'letters') {
-        // Letras con tildes, diéresis, ñ, Ñ, mayúsculas, minúsculas, espacios y guiones
         finalValue = finalValue.replace(/[^a-zA-ZÁÉÍÓÚÜÑáéíóúüñ\s\-]/g, '')
     } else if (props.inputRestrict === 'numbers') {
-        // Números, punto decimal, comas, espacios y guiones (para negativos o formatos)
-        finalValue = finalValue.replace(/[^0-9.,\s\-]/g, '')
+        finalValue = finalValue.replace(/[^0-9]/g, '')
     } else if (props.inputRestrict === 'alphanumeric') {
-        // Letras, números, tildes, ñ, espacios, guiones, guion bajo, punto y coma
-        finalValue = finalValue.replace(/[^a-zA-Z0-9ÁÉÍÓÚÜÑáéíóúüñ\s\-_.,;]/g, '')
-    }else{
-        // Letras, números, tildes, ñ, espacios, guiones, guion bajo, punto y coma
         finalValue = finalValue.replace(/[^a-zA-Z0-9ÁÉÍÓÚÜÑáéíóúüñ\s\-_.,;]/g, '')
     }
 
@@ -187,6 +195,21 @@ function updateValue(val) {
     emit('update:modelValue', finalValue)
 }
 
+function onKeydown(event) {
+    if (props.inputRestrict === 'letters') {
+        if (!/[a-zA-ZÁÉÍÓÚÜÑáéíóúüñ\s\-]/.test(event.key) && event.key !== 'Backspace') {
+            event.preventDefault()
+        }
+    } else if (props.inputRestrict === 'numbers') {
+        if (!/[0-9]/.test(event.key) && event.key !== 'Backspace') {
+            event.preventDefault()
+        }
+    } else if (props.inputRestrict === 'alphanumeric') {
+        if (!/[a-zA-Z0-9ÁÉÍÓÚÜÑáéíóúüñ\s\-_.,;]/.test(event.key) && event.key !== 'Backspace') {
+            event.preventDefault()
+        }
+    }
+}
 
 function onFocus(e) {
     isFocused.value = true
@@ -195,6 +218,9 @@ function onFocus(e) {
 
 function onBlur(e) {
     isFocused.value = false
+    if (props.regex && internalValue.value && !props.regex.test(internalValue.value)) {
+        internalError.value = 'El formato no es válido'
+    }
     emit('blur', e)
 }
 
@@ -205,24 +231,6 @@ onMounted(() => {
     prefersDark.addEventListener('change', (e) => {
         isDark.value = e.matches
     })
-
-    /* Init mask if provided */
-    if (props.mask && inputRef.value) {
-        maskRef = IMask(
-            inputRef.value,
-            typeof props.mask === 'string' ? { mask: props.mask } : props.mask
-        )
-        maskRef.on('accept', () => {
-            emit('update:modelValue', maskRef.value)
-        })
-    }
-})
-
-/* ======================= Cleanup ========================== */
-onBeforeUnmount(() => {
-    if (maskRef) {
-        maskRef.destroy()
-        maskRef = null
-    }
 })
 </script>
+